requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
colorclass==2.2.2
compressed-rtf==1.0.7
cryptography==45.0.5
defusedxml==0.7.1
distro==1.9.0
easygui==0.98.3
ebcdic==1.1.1
extract-msg==0.54.1
fastapi==0.116.1
filelock==3.18.0
fsspec==2025.7.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpie==3.2.4
httpx==0.28.1
huggingface-hub==0.34.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.72
langchain-text-splitters==0.3.9
langsmith==0.4.8
lark==1.1.9
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
msoffcrypto-tool==5.4.2
multidict==6.6.3
networkx==3.5
numpy==2.3.2
olefile==0.47
oletools==0.60.2
ollama==0.5.1
openai==1.97.1
orjson==3.11.1
packaging==24.2
pandas==2.3.1
pcodedmp==1.2.6
pillow==11.3.0
pinecone==7.3.0
pinecone-plugin-assistant==1.7.0
pinecone-plugin-interface==0.0.7
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyMuPDF==1.26.3
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
red-black-tree-mod==1.22
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.1.0
RTFDE==0.1.2.1
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.0
sentence-transformers==5.0.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.47.2
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.54.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
wheel==0.45.1
win_unicode_console==0.5
zstandard==0.23.0
